% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{dfr_wei}
\alias{dfr_wei}
\alias{qfr_wei}
\alias{pfr_wei}
\title{The Frechet-Weibull Distribution}
\usage{
dfr_wei(x, fr_shape = 2, fr_scale = 100, wei_shape = 2,
  wei_scale = 1000, fr_weight = 0.5, log = FALSE)

qfr_wei(p, fr_shape = 2, fr_scale = 100, wei_shape = 2,
  wei_scale = 1000, fr_weight = 0.5, log = FALSE)

pfr_wei(q, fr_shape = 2, fr_scale = 100, wei_shape = 2,
  wei_scale = 1000, fr_weight = 0.5, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{fr_shape}{the Frechet shape}

\item{fr_scale}{the Frechet scale}

\item{wei_shape}{the Weibull shape}

\item{wei_scale}{the Weibull scale}

\item{fr_weight}{numeric value from 0 to 1, the Frechet weight, where
the corresponding Weibull weight is \code{(1 - fr_weight)}.}

\item{log}{logical passed to \code{actuar::dinvweibull}, if \code{TRUE}
probabilities/densities are returned as \code{log(p)}.}
}
\description{
The Frechet-Weibull Distribution
}
\details{
Density, distribution function, quantile function,
and random generation for the combined Frechet-Weibull
distribution with five parameters: \code{fr_shape}, \code{fr_scale},
\code{wei_shape}, \code{wei_scale}, and \code{fr_weight}.
}
\seealso{
Other SALSA distribution fit functions: \code{\link{fitdist_fr_wei}},
  \code{\link{fitdist_fr}}, \code{\link{params_fr_wei}}
}
\concept{SALSA distribution fit functions}
