% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{fitdist_fr}
\alias{fitdist_fr}
\title{Fit Frechet distribution to non-censored data}
\usage{
fitdist_fr(x, method = "mme", min_shape = 1, order = c(1, 2),
  central = FALSE, absolute = FALSE, na.rm = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{x}{numeric vector}

\item{method}{fitting method, argument passed to
\code{fitdistrplus::fitdist()}, by default \code{"mme"} moment
matching estimation.}

\item{min_shape}{numeric value restricting the minimum shape
otherwise output is set to \code{NULL}.}

\item{order, central, absolute, na.rm}{arguments passed to
\code{moments::moment()}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are passed to
\code{fitdistrplus::fitdist()}.}

\item{param_fr_wei}{\code{data.frame} of parameter bounds, output
from \code{params_fr_wei()}, which defines the start, upper, and
lower bounds for each of the five parameters.}
}
\value{
object of class \code{"fitdist"} described in
\code{fitdistrplus::fitdist()}. If any error occurs during
the fit, a \code{NULL} is returned.
}
\description{
Fit Frechet distribution to non-censored data
}
\details{
This function is a simple wrapper around \code{fitdistrplus::fitdist()}
for the Frechet distribution, using \code{"actuar::dinvweibull"}.
Most customizations are performed by the corresponding
support function, for example when \code{method="mme"} the
distribution fit function is \code{fitdistrplus::mmedist()},
which then calls \code{stats::optim()}. Arguments are
passed to these functions using \code{...} where applicable.
}
\examples{
data(oz2_numi_per_cell);
usecounts <- oz2_numi_per_cell$count[oz2_numi_per_cell$count >= 16];
fit_fr <- fitdist_fr(x=usecounts);
coef(fit_fr);

}
