% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{fitdist_fr_wei}
\alias{fitdist_fr_wei}
\title{Fit Frechet-Weibull distribution to non-censored data}
\usage{
fitdist_fr_wei(x, param_fr_wei, method = "mle",
  optim.method = "Nelder-Mead", log = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{x}{numeric vector}

\item{param_fr_wei}{\code{data.frame} of parameter bounds, output
from \code{params_fr_wei()}, which defines the start, upper, and
lower bounds for each of the five parameters.}

\item{method}{fitting method, argument passed to
\code{fitdistrplus::fitdist()}.}

\item{optim.method}{character \code{method} argument passed to
\code{fitdistrplus::mledist} or corresponding function based
upon the \code{method} argument, but ultimately passed to
\code{stats::optim()}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are passed to
\code{fitdistrplus::fitdist()}.}
}
\value{
object of class \code{"fitdist"} described in
\code{fitdistrplus::fitdist()}.
}
\description{
Fit Frechet-Weibull distribution to non-censored data
}
\details{
This function is a simple wrapper around \code{fitdistrplus::fitdist()}
for the five-parameter Frechet-Weibull distribution.
Most customizations are performed by the corresponding
support function, for example when \code{method="mle"} the
distribution fit function is \code{fitdistrplus::mledist()},
which then calls \code{stats::optim()}. Arguments are
passed to these functions using \code{...} where applicable.
}
