% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{get_lower_bound}
\alias{get_lower_bound}
\title{Calculate lower bound from Frechet-Weibull fit}
\usage{
get_lower_bound(fit_fr_wei, ...)
}
\arguments{
\item{fit_fr_wei}{\code{fitdist} object output from \code{fitdist_fr_wei()},
or numeric vector with names \code{"fr_weight"}, \code{"fr_shape"},
\code{"fr_scale"}, \code{"wei_shape"}, \code{"wei_scale"}.}

\item{...}{additional arguments are ignored.}
}
\description{
Calculate lower bound from Frechet-Weibull fit
}
