% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{get_salsa_steps}
\alias{get_salsa_steps}
\title{Get step parameters for SALSA}
\usage{
get_salsa_steps(x, include_vector = FALSE, ...)
}
\arguments{
\item{x}{numeric vector of counts}

\item{...}{additional arguments are ignored}
}
\description{
Get step parameters for SALSA
}
\details{
This function takes a vector of counts and determines the
appropriate step size to use when iterating the count
threshold used by SALSA.
}
\examples{
library(salsa);
data(oz2_numi_per_cell);

usecounts <- sort(oz2_numi_per_cell$count);
get_salsa_steps(usecounts);

# optionally return the vector of thresholds to use
get_salsa_steps(usecounts, include_vector=TRUE);

}
