% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{get_salsa_table}
\alias{get_salsa_table}
\title{Get SALSA results as a data.frame}
\usage{
get_salsa_table(fit_list, verbose = FALSE, ...)
}
\arguments{
\item{fit_list}{\code{list} output from \code{do_salsa_steps()},
where each list element contains \code{min_count},
and one or more of \code{fit_fr} and \code{fit_fr_wei}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} containing one row per minimum count
threshold, with columns containing the Frechet and
Frechet-Weibull fit parameters, or NA values when there
are no parameters for the given minimum count threshold.
}
\description{
Get SALSA results as a data.frame
}
\details{
This function converts the list of lists output from
\code{do_salsa_steps()} into a "tidy" data.frame intended
to be used for visualization.
}
\examples{
library(salsa);
data(oz2_numi_per_cell);
x <- oz2_numi_per_cell$count[oz2_numi_per_cell$count >= 16];
x_salsa <- do_salsa_steps(x,
   count_vector=c(16,32,128),
   cache_fr=NULL,
   cache_fr_wei=NULL);
x_df <- get_salsa_table(x_salsa);
x_df;

}
\seealso{
Other SALSA core functions: \code{\link{do_salsa_steps}}
}
\concept{SALSA core functions}
