% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{get_upper_bound}
\alias{get_upper_bound}
\title{Calculate upper bound from Frechet fit}
\usage{
get_upper_bound(fit_fr, ...)
}
\arguments{
\item{fit_fr}{\code{fitdist} object output from \code{fitdist_fr()},
or numeric vector with names \code{"shape"}, \code{"scale"}.}

\item{...}{additional arguments are ignored.}
}
\description{
Calculate upper bound from Frechet fit
}
