% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-shiny.R
\name{launchSalsaApp}
\alias{launchSalsaApp}
\title{Launch SALSA R-Shiny app}
\usage{
launchSalsaApp(...)
}
\arguments{
\item{...}{additional arguments are ignored.}
}
\description{
Launch SALSA R-Shiny app
}
\details{
This function launches the SALSA workflow as an
R-shiny app.
}
\examples{
# Note: disabled for web page examples
# launchSalsaApp();

}
\seealso{
Other SALSA Shiny functions: \code{\link{salsaAppConstants}},
  \code{\link{salsaAppServer}}, \code{\link{salsaAppUI}}
}
\concept{SALSA Shiny functions}
