% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-data.R
\docType{data}
\name{oz2_numi_per_cell}
\alias{oz2_numi_per_cell}
\title{Sample data scRNA-seq nUMI per Cell}
\format{\code{data.frame} with two colnames:
\itemize{
\item \code{barcode} represents the unique sequence barcode
for a given cell.
\item \code{count} represents the total number of UMI
per cell from the raw data, without filtering.
}}
\usage{
oz2_numi_per_cell
}
\description{
Sample data scRNA-seq nUMI per Cell
}
\seealso{
Other SALSA example data: \code{\link{oz2_numi_per_gene}}
}
\concept{SALSA example data}
\keyword{datasets}
