% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-data.R
\docType{data}
\name{oz2_numi_per_gene}
\alias{oz2_numi_per_gene}
\title{Sample data scRNA-seq nUMI per Gene}
\format{\code{data.frame} with two colnames:
\itemize{
\item \code{gene} represents the annotated gene symbol.
\item \code{count} represents the total number of UMI
per gene from the raw data, without filtering.
}}
\usage{
oz2_numi_per_gene
}
\description{
Sample data scRNA-seq nUMI per Gene
}
\seealso{
Other SALSA example data: \code{\link{oz2_numi_per_cell}}
}
\concept{SALSA example data}
\keyword{datasets}
