% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{params_fr_wei}
\alias{params_fr_wei}
\title{Parameter bounds for Frechet-Weibull fit}
\usage{
params_fr_wei(param_fr_wei = NULL, ...)
}
\arguments{
\item{param_fr_wei}{\code{data.frame} containing colnames
\code{"fr_shape"}, \code{"fr_scale"}, \code{"wei_shape"}, \code{"wei_scale"},
and \code{"fr_weight"}, and rownames \code{"start"}, \code{"lower"},
\code{"upper"}.}

\item{...}{additional arguments are ignored.}
}
\description{
Parameter bounds for Frechet-Weibull fit
}
\details{
This function defines a data.frame of parameter bounds,
restricting the range of values allowed during the
Frechet-Weibull distribution fit. Each of the five
parameters are given \code{start}, \code{lower}, and \code{upper}
values:
\itemize{
\item \code{fr_weight} the weight of Frechet relative to Weibull,
on a scale of 0 to 1.
\item \code{fr_shape} the shape value for the Frechet distribution
\item \code{fr_scale} the scale value for the Frechet distribution
\item \code{wei_shape} the shape value for the Weibull distribution
\item \code{wei_scale} the scale value for the Weibull distribution
}
}
