% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-plot.R
\name{rank_count_plot}
\alias{rank_count_plot}
\title{Rank-Count plot}
\usage{
rank_count_plot(x, UMIname = "UMIname", log = c("x", "y"),
  flip_axes = FALSE, highlight_range = NULL, ...)
}
\arguments{
\item{x}{numeric or integer vector of counts}

\item{...}{additional arguments are ignored}
}
\description{
Rank-Count plot
}
\details{
This function creates a classic "knee plot" showing
\code{"Rank"} on the x-axis, and \code{"Count"} on the y-axis,
typically with log-scaled x- and y-axes.
}
