% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{read_numi}
\alias{read_numi}
\title{Read nUMI tab-delimited file}
\usage{
read_numi(file = NULL, text = NULL, sep = "", ...)
}
\arguments{
\item{file}{input file connection}

\item{text}{optional text input to \code{utils::read.table()}, used when
there is no \code{file}.}

\item{...}{additional arguments are passed to \code{utils::read.table()}.}
}
\value{
\code{data.frame} or NULL if validation checks fail.
}
\description{
Read nUMI tab-delimited file
}
\details{
This function is a simple wrapper around \code{utils::read.table()}
with some follow-up data cleaning to handle optional header
row, and sometimes having space delimiter instead of tab.

For most validation checks, NULL is returned. Data is expected to
have at least 2 columns, where column 2 has class \code{"numeric"}
or \code{"integer"}.
}
\examples{
library(salsa);
data(oz2_numi_per_cell);
text <- jamba::pasteByRow(oz2_numi_per_cell[1:10,], sep=" ");
read_numi(text=text);

}
