% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-shiny-constants.R
\name{salsaAppConstants}
\alias{salsaAppConstants}
\title{SALSA Shiny app constants}
\usage{
salsaAppConstants(...)
}
\description{
SALSA Shiny app constants
}
\details{
This function is intended to define constant values used in
the creation of the SALSA shiny UI.

This function is intended to be called only from within
an R-shiny app, since it defines several variables in the
parent (global) environment.
}
\seealso{
Other SALSA Shiny functions: \code{\link{launchSalsaApp}},
  \code{\link{salsaAppServer}}, \code{\link{salsaAppUI}}
}
\concept{SALSA Shiny functions}
