% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-shiny-server.R
\name{salsaAppServer}
\alias{salsaAppServer}
\title{SALSA Shiny app server}
\usage{
salsaAppServer(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}
}
\description{
SALSA Shiny app sserver
}
\seealso{
Other SALSA Shiny functions: \code{\link{launchSalsaApp}},
  \code{\link{salsaAppConstants}}, \code{\link{salsaAppUI}}
}
\concept{SALSA Shiny functions}
