% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-shiny-ui.R
\name{salsaAppUI}
\alias{salsaAppUI}
\title{SALSA Shiny app UI}
\usage{
salsaAppUI(...)
}
\description{
SALSA Shiny app UI
}
\seealso{
Other SALSA Shiny functions: \code{\link{launchSalsaApp}},
  \code{\link{salsaAppConstants}},
  \code{\link{salsaAppServer}}
}
\concept{SALSA Shiny functions}
