% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{do_salsa_steps}
\alias{do_salsa_steps}
\title{Perform SALSA steps for threshold detection}
\usage{
do_salsa_steps(x, n_vector = NULL, n_start = NULL, n_bound = NULL,
  step_size = NULL, ...)
}
\description{
Perform SALSA steps for threshold detection
}
