% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{fitdist_fr_wei}
\alias{fitdist_fr_wei}
\title{Fit Frechet-Weibull distribution to non-censored data}
\usage{
fitdist_fr_wei(x, param_fr_wei = params_fr_wei(), method = "mle",
  optim.method = "Nelder-Mead", log = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{x}{numeric vector}

\item{param_fr_wei}{\code{data.frame} of parameter bounds, output
from \code{params_fr_wei()}, which defines the start, upper, and
lower bounds for each of the five parameters.}

\item{method}{fitting method, argument passed to
\code{fitdistrplus::fitdist()}.}

\item{optim.method}{character \code{method} argument passed to
\code{fitdistrplus::mledist} or corresponding function based
upon the \code{method} argument, but ultimately passed to
\code{stats::optim()}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are passed to
\code{fitdistrplus::fitdist()}.}
}
\value{
object of class \code{"fitdist"} described in
\code{fitdistrplus::fitdist()}.
}
\description{
Fit Frechet-Weibull distribution to non-censored data
}
\details{
This function is a simple wrapper around \code{fitdistrplus::fitdist()}
for the five-parameter Frechet-Weibull distribution.
Most customizations are performed by the corresponding
support function, for example when \code{method="mle"} the
distribution fit function is \code{fitdistrplus::mledist()},
which then calls \code{stats::optim()}. Arguments are
passed to these functions using \code{...} where applicable.

TODO: This function currently expects to use \code{method="mle"}
which utilizes the upper,lower,start parameters for each
of the five parameters. Other methods may have different
requirements which are not formally handled. In future
we will evaluate and support additional methods, modifying
the call to \code{fitdistrplus::fitdist()} as needed.
}
\examples{
library(salsa);
data(oz2_numi_per_cell);
x <- oz2_numi_per_cell$count[oz2_numi_per_cell$count >= 16];
fit_fr_wei <- fitdist_fr_wei(x);
# print coefficients
print(coef(fit_fr_wei));

}
\seealso{
Other SALSA distribution fit functions: \code{\link{dfr_wei}},
  \code{\link{fitdist_fr}}, \code{\link{params_fr_wei}}
}
\concept{SALSA distribution fit functions}
