% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{get_lower_bound}
\alias{get_lower_bound}
\title{Calculate lower bound from Frechet-Weibull fit}
\usage{
get_lower_bound(fit_fr_wei, ...)
}
\arguments{
\item{fit_fr_wei}{\code{fitdist} object output from \code{fitdist_fr_wei()},
or numeric vector with names \code{"fr_weight"}, \code{"fr_shape"},
\code{"fr_scale"}, \code{"wei_shape"}, \code{"wei_scale"}.}

\item{...}{additional arguments are ignored.}
}
\value{
numeric value representing the lowest number of UMI to
accept for downstream analysis of scRNA-seq data.
}
\description{
Calculate lower bound from Frechet-Weibull fit
}
\details{
This function takes the output from \code{fitdist_fr_wei()} and
calculates the associated predicted lower bound for
number of UMI to accept for downstream analysis of
scRNA-seq data.

The equation:

fr_scale * (wei_scale / fr_scale) ^ (fr_weight / (fr_shape * wei_shape))
}
\seealso{
Other SALSA support functions: \code{\link{get_salsa_steps}},
  \code{\link{get_upper_bound}}
}
\concept{SALSA support functions}
