% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salsa-fit.R
\name{plot_fr_wei}
\alias{plot_fr_wei}
\title{Plot Frechet-Weibull fit}
\usage{
plot_fr_wei(fit, addx = NULL, ylim = NULL, xlim = NULL,
  scale = FALSE, log = "x", type = c("gg", "base"), ...)
}
\description{
Plot Frechet-Weibull fit
}
